+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/HttpGoogleProjectDAO.scala
@@ -11,6 +11,7 @@ import com.google.api.services.servicemanagement.ServiceManagement
 import com.google.api.services.servicemanagement.model.EnableServiceRequest
 import org.broadinstitute.dsde.workbench.google.GoogleCredentialModes._
 import org.broadinstitute.dsde.workbench.google.GoogleCredentialModes.GoogleCredentialMode
+import org.broadinstitute.dsde.workbench.google.GoogleUtilities.RetryPredicates._
 import org.broadinstitute.dsde.workbench.metrics.GoogleInstrumentedService
 import org.broadinstitute.dsde.workbench.model.google.GoogleResourceTypes.GoogleParentResourceType
 
@@ -40,7 +41,7 @@ class HttpGoogleProjectDAO(appName: String,
   }
 
   override def createProject(projectName: String): Future[String] = {
-    retryWhen500orGoogleError(() => {
+    retry(when5xx, whenUsageLimited, when404, whenInvalidValueOnBucketCreation, whenNonHttpIOException)(() => {
       executeGoogleRequest(cloudResManager.projects().create(new Project().setName(projectName).setProjectId(projectName)))
     }).map { operation =>
       operation.getName
@@ -48,7 +49,7 @@ class HttpGoogleProjectDAO(appName: String,
   }
 
   override def createProject(projectName: String, parentId: String, parentType: GoogleParentResourceType): Future[String] = {
-    retryWhen500orGoogleError(() => {
+    retry(when5xx, whenUsageLimited, when404, whenInvalidValueOnBucketCreation, whenNonHttpIOException)(() => {
       executeGoogleRequest(cloudResManager.projects().create(new Project().setName(projectName).setProjectId(projectName)
         .setParent(new ResourceId().setId(parentId).setType(parentType.value))))
     }).map { operation =>
@@ -57,13 +58,13 @@ class HttpGoogleProjectDAO(appName: String,
   }
 
   override def pollOperation(operationId: String): Future[Operation] = {
-    retryWhen500orGoogleError(() => {
+    retry(when5xx, whenUsageLimited, when404, whenInvalidValueOnBucketCreation, whenNonHttpIOException)(() => {
       executeGoogleRequest(cloudResManager.operations().get(operationId))
     })
   }
 
   override def isProjectActive(projectName: String): Future[Boolean] = {
-    retryWithRecoverWhen500orGoogleError { () =>
+    retryWithRecover(when5xx, whenUsageLimited, when404, whenInvalidValueOnBucketCreation, whenNonHttpIOException) { () =>
       // get the project
       Option(executeGoogleRequest(cloudResManager.projects().get(projectName)))
     } {
@@ -78,7 +79,7 @@ class HttpGoogleProjectDAO(appName: String,
   }
 
   override def isBillingActive(projectName: String): Future[Boolean] = {
-    retryWithRecoverWhen500orGoogleError { () =>
+    retryWithRecover(when5xx, whenUsageLimited, when404, whenInvalidValueOnBucketCreation, whenNonHttpIOException) { () =>
       Option(executeGoogleRequest(billing.projects().getBillingInfo(s"projects/$projectName")))
     } {
       // if the project doesn't exist, don't fail
@@ -91,7 +92,7 @@ class HttpGoogleProjectDAO(appName: String,
   }
 
   override def enableService(projectName: String, serviceName: String): Future[String] = {
-    retryWhen500orGoogleError(() => {
+    retry(when5xx, whenUsageLimited, when404, whenInvalidValueOnBucketCreation, whenNonHttpIOException)(() => {
       executeGoogleRequest(serviceManagement.services().enable(serviceName, new EnableServiceRequest().setConsumerId(s"project:$projectName")))
     }).map { operation =>
       operation.getName
@@ -99,7 +100,7 @@ class HttpGoogleProjectDAO(appName: String,
   }
 
   override def getLabels(projectName: String): Future[Map[String, String]] = {
-    retryWhen500orGoogleError { () =>
+    retry(when5xx, whenUsageLimited, when404, whenInvalidValueOnBucketCreation, whenNonHttpIOException) { () =>
       // get the project
       executeGoogleRequest(cloudResManager.projects().get(projectName))
     } map { project =>
@@ -108,7 +109,7 @@ class HttpGoogleProjectDAO(appName: String,
   }
 
   override def getAncestry(projectName: String): Future[Seq[Ancestor]] = {
-    retryWhen500orGoogleError(() => {
+    retry(when5xx, whenUsageLimited, when404, whenInvalidValueOnBucketCreation, whenNonHttpIOException)(() => {
       executeGoogleRequest(cloudResManager.projects().getAncestry(projectName, new GetAncestryRequest()))
     }).map { ancestry =>
       Option(ancestry.getAncestor).map(_.asScala).getOrElse(Seq.empty)